+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Editable tables no longer show their heading row if there aren’t any data rows yet. ([#7158](https://github.com/craftcms/cms/issues/7158))
 - Relational fields’ “Show the site menu” settings are no longer enabled by default.
 - Relational fields’ element selection modals now always default to the source element’s site, if no target site is specified in the field’s settings. ([#7164](https://github.com/craftcms/cms/issues/7164))
+- Controllers now run request and permission checks from `beforeAction()` rather than `init()`. ([#7168](https://github.com/craftcms/cms/issues/7168))
 
 ### Deprecated
 - Deprecated `craft\fields\BaseRelationField::inputSiteId()`.